ml/requirements.txt
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
fastapi==0.103.1
uvicorn==0.23.2
pydantic==1.10.13
python-dotenv==1.0.0
gunicorn==21.2.0 